+++ src/js/const/asset.const.js
@@ -2,5 +2,3 @@ export const ASSET_STATES = Object.freeze({
   active: 0,
   deleted: 1,
 })
-
-export const NEW_CREATE_ASSET_REQUEST_ID = '0'